wallet: move legacy helpers to deprecated
