+++ wallet/wallet.go
@@ -9,7 +9,7 @@
 //
 // TODO(yy): bring wrapcheck back when implementing the `Store` interface.
 //
-//nolint:wrapcheck,cyclop
+//nolint:wrapcheck
 package wallet
 
 import (
@@ -19,9 +19,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/btcsuite/btcd/btcutil/hdkeychain"
 	"github.com/btcsuite/btcd/chaincfg"
-	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcwallet/chain"
 	"github.com/btcsuite/btcwallet/waddrmgr"
 	"github.com/btcsuite/btcwallet/wallet/internal/db"
@@ -410,159 +408,9 @@ func (w *Wallet) ID() uint32 {
 	return w.id
 }
 
-// RemoveDescendants attempts to remove any transaction from the wallet's tx
-// store (that may be unconfirmed) that spends outputs created by the passed
-// transaction. This remove propagates recursively down the chain of descendent
-// transactions.
-func (w *Wallet) RemoveDescendants(tx *wire.MsgTx) error {
-	txRecord, err := wtxmgr.NewTxRecordFromMsgTx(tx, time.Now())
-	if err != nil {
-		return err
-	}
-
-	return walletdb.Update(w.db, func(tx walletdb.ReadWriteTx) error {
-		wtxmgrNs := tx.ReadWriteBucket(wtxmgrNamespaceKey)
-
-		return w.txStore.RemoveUnminedTx(wtxmgrNs, txRecord)
-	})
-}
-
-// BirthdayBlock returns the birthday block of the wallet.
-//
-// NOTE: The wallet won't start until the backend is synced, thus the birthday
-// block won't be set and `ErrBirthdayBlockNotSet` will be returned.
-func (w *Wallet) BirthdayBlock() (*waddrmgr.BlockStamp, error) {
-	var birthdayBlock waddrmgr.BlockStamp
-
-	// Query the wallet's birthday block height from db.
-	err := walletdb.View(w.db, func(tx walletdb.ReadTx) error {
-		addrmgrNs := tx.ReadBucket(waddrmgrNamespaceKey)
-
-		bb, _, err := w.addrStore.BirthdayBlock(addrmgrNs)
-		birthdayBlock = bb
-
-		return err
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return &birthdayBlock, nil
-}
-
 // SyncedTo calls the `SyncedTo` method on the wallet's manager.
 func (w *Wallet) SyncedTo() waddrmgr.BlockStamp {
 	return w.addrStore.SyncedTo()
-}
-
-// AddrManager returns the internal address manager.
-//
-// TODO(yy): Refactor it in lnd and remove the method.
-func (w *Wallet) AddrManager() waddrmgr.AddrStore {
-	return w.addrStore
-}
-
-// NotificationServer returns the internal NotificationServer.
-//
-// TODO(yy): Refactor it in lnd and remove the method.
-func (w *Wallet) NotificationServer() *NotificationServer {
-	return w.NtfnServer
-}
-
-// CreateWithCallback is the same as Create with an added callback that will be
-// called in the same transaction the wallet structure is initialized.
-func CreateWithCallback(db walletdb.DB, pubPass, privPass []byte,
-	rootKey *hdkeychain.ExtendedKey, params *chaincfg.Params,
-	birthday time.Time, cb func(walletdb.ReadWriteTx) error) error {
-
-	return create(
-		db, pubPass, privPass, rootKey, params, birthday, false, cb,
-	)
-}
-
-// CreateWatchingOnlyWithCallback is the same as CreateWatchingOnly with an
-// added callback that will be called in the same transaction the wallet
-// structure is initialized.
-func CreateWatchingOnlyWithCallback(db walletdb.DB, pubPass []byte,
-	params *chaincfg.Params, birthday time.Time,
-	cb func(walletdb.ReadWriteTx) error) error {
-
-	return create(
-		db, pubPass, nil, nil, params, birthday, true, cb,
-	)
-}
-
-// CreateWatchingOnly creates an new watch-only wallet, writing it to
-// an empty database. No root key can be provided as this wallet will be
-// watching only.  Likewise no private passphrase may be provided
-// either.
-func CreateWatchingOnly(db walletdb.DB, pubPass []byte,
-	params *chaincfg.Params, birthday time.Time) error {
-
-	return create(
-		db, pubPass, nil, nil, params, birthday, true, nil,
-	)
-}
-
-func create(db walletdb.DB, pubPass, privPass []byte,
-	rootKey *hdkeychain.ExtendedKey, params *chaincfg.Params,
-	birthday time.Time, isWatchingOnly bool,
-	cb func(walletdb.ReadWriteTx) error) error {
-
-	// If no root key was provided, we create one now from a random seed.
-	// But only if this is not a watching-only wallet where the accounts are
-	// created individually from their xpubs.
-	if !isWatchingOnly && rootKey == nil {
-		hdSeed, err := hdkeychain.GenerateSeed(
-			hdkeychain.RecommendedSeedLen,
-		)
-		if err != nil {
-			return err
-		}
-
-		// Derive the master extended key from the seed.
-		rootKey, err = hdkeychain.NewMaster(hdSeed, params)
-		if err != nil {
-			return fmt.Errorf("failed to derive master extended " +
-				"key")
-		}
-	}
-
-	// We need a private key if this isn't a watching only wallet.
-	if !isWatchingOnly && rootKey != nil && !rootKey.IsPrivate() {
-		return fmt.Errorf("need extended private key for wallet that " +
-			"is not watching only")
-	}
-
-	return walletdb.Update(db, func(tx walletdb.ReadWriteTx) error {
-		addrmgrNs, err := tx.CreateTopLevelBucket(waddrmgrNamespaceKey)
-		if err != nil {
-			return err
-		}
-		txmgrNs, err := tx.CreateTopLevelBucket(wtxmgrNamespaceKey)
-		if err != nil {
-			return err
-		}
-
-		err = waddrmgr.Create(
-			addrmgrNs, rootKey, pubPass, privPass, params, nil,
-			birthday,
-		)
-		if err != nil {
-			return err
-		}
-
-		err = wtxmgr.Create(txmgrNs)
-		if err != nil {
-			return err
-		}
-
-		if cb != nil {
-			return cb(tx)
-		}
-
-		return nil
-	})
 }
 
 // hasMinConfs checks whether a transaction at height txHeight has met minconf
